# packages/repograph/docs/architecture/api-contract.yaml
openapi: 3.0.3
info:
  title: RepoGraph API
  version: 1.0.0
  description: |
    RepoGraph API for code intelligence and AI-readiness.

    **Stage 1 Features:**
    - Repository indexing (SCIP + fallback)
    - Graph queries (search, ego, impact)
    - AI-readiness scorecard
    - Safe autofixers
    - Guardrails and telemetry

    **Authentication:**
    - JWT Bearer tokens (OIDC)
    - API keys (X-API-Key header)

    **Multi-tenancy:**
    - All resources scoped to organizations
    - Automatic tenant isolation via RLS
  contact:
    name: RepoGraph Team
    email: repograph@aeptus.com
  license:
    name: Proprietary
servers:
  - url: https://repograph.aeptus.com/api/v1
    description: Production
  - url: https://staging.repograph.aeptus.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user and organization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, org_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                org_name:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from OIDC provider

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
...
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        org_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, member, readonly]
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...